requirements.txt
fastapi~=0.103.0
joblib~=1.4.2
pandas~=2.2.2
catboost~=1.2.5
numpy~=1.26.4
matplotlib~=3.9.0
scikit-learn~=1.4.2
shap~=0.42.1
streamlit~=1.36.0
uvicorn~=0.30.1
pydantic~=2.5.3
loguru~=0.5.3
halo~=0.0.31
gradio~=4.36.1
seaborn~=0.13.2
plotly~=5.22.0
pillow~=10.3.0